requirements.txt
bcrypt==4.1.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
protobuf==4.21.12
Pygments==2.19.2
PyJWT==2.8.0
pytest==8.4.2
python-dotenv==1.0.0
requests==2.32.5
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.0.0
zipp==3.23.0
